fix: Detect DPP & DPP broadcast support with __GFX<GENERATION>__ macros
The amdgpu target in clang now provides the GFX generation as a
predefined macro, so we no longer need to explicitly list all targets,
which was bad for maintenance.

Also replace the use of the generic `ROCPRIM_NAVI` which signals navi support,
with `ROCPRIM_DETAIL_HAS_DPP_BROADCAST`, a macro that explicitly
states what we're after.

Also also makes sure that `ROCPRIM_DETAIL_USE_DPP` is always defined
(to 0 when DPP is disabled), previously it was undefined when
`ROCPRIM_DISABLE_DPP` was set.